Fix: Combining header and header Nav
